% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAI-fun2.R
\name{align}
\alias{align}
\title{Spectral Manifold Alignment and Inference (SMAI)}
\usage{
align(
  data1,
  data2,
  sel1 = NULL,
  sel2 = NULL,
  r.max = 200,
  t = 5,
  test = TRUE,
  dir.map = c("1to2", "2to1", "auto"),
  denoise = c("scree", "screeNOT", "none"),
  outlier.cut = 20,
  knn = 30,
  prop.align = 0.3,
  prop.inf = 0.5,
  cutoff = 1.001,
  cutoff.t = 1.5
)
}
\arguments{
\item{data1}{data matrix (pxn1).}

\item{data2}{data matrix (pxn2).}

\item{sel1}{index set to select a subset of data1 for SMAI-align}

\item{sel2}{index set to select a subset of data2 for SMAI-align}

\item{r.max}{intetger (default 50); the maximum number of singular values selected for desnoising and SMAI-alignment.}

\item{t}{integer (default 5); maximum number of iterations for SMAI-align.}

\item{test}{logical (default TRUE); whether SMAI-test is included in the workflow.}

\item{dir.map}{directionality of alignment; can be data1 aligned to data2 ("1to2"), data2 aligned to data1 ("2to1"), or internally determined acording to alignment error ("auto").}

\item{denoise}{whether data should be denoised before searching for the spectral alignment; "scree" uses scree plot to denoise; "screeNOT" uses the ScreeNOT package to denoise; "none" does not perform denoising to the data.}

\item{outlier.cut}{integer (default 20); the number of outliers to be removed from the reference data at each iteration.}

\item{knn}{integer (default 50); the number of nearest neighbors to be sampled from to do SMAI-test. A smaller value leads to more stringent selection of the test set.}

\item{prop.align}{numeric between 0 and 1 (default 0.5; and must not exceed 0.5 if test=TRUE, to ensure); proportion of samples used for SMAI-alignment. Smaller value improves speed with a possible cost of precision.}

\item{prop.inf}{numeric between 0 and 1-prop.align (default 0.5); proportion of samples used for SMAI-test. Smaller value improves speed with a possible cost of precision.}

\item{cutoff}{numeric larger than 1 (default 1.05); the threshold value for selection of number of informative singular values in SMAI-alignment. Effective only when denoise="scree."}

\item{cutoff.t}{numeric larger than 1 (default 1.5); the threshold value for data-driven selection of number of informative singular values in SMAI-test.}
}
\value{
data1.integrate: integrated data matrices (pxn1);

data2.integrate: integrated data matrices (pxn2);

feature.rot: rotation matrix;

gamma: mean-shift vector;

beta: scaling factor;

p.value: p-value from alignability testing.
}
\description{
This function enables principled and interpretable alignability testing and structure-preserving integration of single-cell data.
}
