% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAI-fun2.R
\name{assess}
\alias{assess}
\title{Batch evaluation metrics}
\usage{
assess(data, meta_data, straggler = NULL, cutoff = 1.05)
}
\arguments{
\item{data}{integrated data matrix (pxn).}

\item{meta_data}{label information.}

\item{straggler}{labels of samples with non-overlapping class labels.}

\item{cutoff}{cutoff parameter for PCA dimension reduction.}
}
\value{
sihouette.batch: 1 - mean sihouette index; ch.batch: inverse CH index; db.batch: DB index.
}
\description{
This function include a few metrics for evaluating batch effect removal.
}
