% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAI-fun2.R
\name{faithful}
\alias{faithful}
\title{Structure-preserving metrics}
\usage{
faithful(data.registered, data)
}
\arguments{
\item{data.registered}{integrated data matrix (nxp).}

\item{data}{original data matrix (nxp).}
}
\value{
Kendall's tau ("kendall"), Spearman's rho ("spearman"), and Pearson ("pearson").
}
\description{
This function include a few
}
